inference/requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
gitpython